Following paper...

+++ losses.py
@@ -19,16 +19,16 @@ def LossE(device, image, target):
 def LossP(vgg, device, image, target):
     criterion = nn.MSELoss()
     criterion.cuda()
-    vgg_2 = vgg[0]
-    vgg_5 = vgg[1]
+    vgg_2 = vgg[0].float()
+    vgg_5 = vgg[1].float()
     vgg_2.cuda()
     vgg_5.cuda()
 
-    extr_feat = vgg_2(image)
-    real_feat = vgg_2(target)
+    extr_feat = vgg_2(image.float())
+    real_feat = vgg_2(target.float())
     loss_2 = criterion(extr_feat, real_feat.detach())
-    extr_feat = vgg_5(image)
-    real_feat = vgg_5(target)
+    extr_feat = vgg_5(image.float())
+    real_feat = vgg_5(target.float())
     loss_5 = criterion(extr_feat, real_feat.detach())
 
     return (2e-1)*loss_2 + (2e-2)*loss_5
@@ -36,7 +36,7 @@ def LossP(vgg, device, image, target):
 
 """ GAN generator and discriminator Losses """
 def LossA(generator, discriminator, device, image, target, bicub, optim_d, lossT=False):
-    disc_train_real = target.to(device)
+    disc_train_real = target.float().to(device)
     batch_size = disc_train_real.size(0)
 
     if lossT:
@@ -46,16 +46,16 @@ def LossA(generator, discriminator, device, image, target, bicub, optim_d, lossT
     # Discriminator true
     optim_d.zero_grad()
     label = torch.full((batch_size,), random.uniform(0.9, 1.0), device=device).cuda()
-    output_d = discriminator(disc_train_real).view(-1)
-    loss_d_real = criterion(output_d, label).cuda()
+    output_d = discriminator(disc_train_real.float()).view(-1)
+    loss_d_real = criterion(output_d.float(), label.float()).cuda()
     D_x = output_d.mean().item()
     loss_d_real.backward()
 
     # Discriminator false
-    output_g = generator(image, bicub)
+    output_g = generator(image.float(), bicub.float())
     output_d = discriminator(output_g.detach()).view(-1)
     label.fill_(random.uniform(0.0, 0.1))
-    loss_d_fake = criterion(output_d, label).cuda()
+    loss_d_fake = criterion(output_d.float(), label.float()).cuda()
     D_G_z1 = output_d.mean().item()
     loss_d = loss_d_real + loss_d_fake
 
@@ -64,8 +64,8 @@ def LossA(generator, discriminator, device, image, target, bicub, optim_d, lossT
 
     # Generator
     label.fill_(random.uniform(0.9, 1.0))
-    output_d = discriminator(output_g).view(-1)
-    loss_g = criterion(output_d, label).cuda()
+    output_d = discriminator(output_g.float()).view(-1)
+    loss_g = criterion(output_d.float(), label.float()).cuda()
     D_G_z2 = output_d.mean().item()
 
     return loss_g, loss_d, D_x, D_G_z1, D_G_z2
@@ -99,9 +99,9 @@ def LossT(vgg, device, image, target, patch_size=16):
     patches_target = torch.cat(patches_target)
     patches_target.to(device)
 
-    loss_1 = criterion(gram_matrix(vgg_1(patches)), gram_matrix(vgg_1(patches_target)))
-    loss_2 = criterion(gram_matrix(vgg_2(patches)), gram_matrix(vgg_2(patches_target)))
-    loss_3 = criterion(gram_matrix(vgg_3(patches)), gram_matrix(vgg_3(patches_target)))
+    loss_1 = criterion(gram_matrix(vgg_1(patches.float())).float(), gram_matrix(vgg_1(patches_target.float())).float())
+    loss_2 = criterion(gram_matrix(vgg_2(patches.float())).float(), gram_matrix(vgg_2(patches_target.float())).float())
+    loss_3 = criterion(gram_matrix(vgg_3(patches.float())).float(), gram_matrix(vgg_3(patches_target.float())).float())
 
     return (3e-7) * loss_1 + (1e-6) * loss_2 + (1e-6) * loss_3
